+++ ghost/core/package.json
@@ -105,7 +105,6 @@
     "@tryghost/members-csv": "0.0.0",
     "@tryghost/members-offers": "0.0.0",
     "@tryghost/metrics": "1.0.37",
-    "@tryghost/milestones": "0.0.0",
     "@tryghost/mw-error-handler": "0.0.0",
     "@tryghost/mw-vhost": "0.0.0",
     "@tryghost/nodemailer": "0.3.45",
