+++ ghost/core/test/unit/server/services/milestones/InMemoryMilestoneRepository.test.js
@@ -1,7 +1,7 @@
 const assert = require('assert/strict');
-const ObjectID = require('bson-objectid');
-const InMemoryMilestoneRepository = require('../lib/InMemoryMilestoneRepository');
-const Milestone = require('../lib/Milestone');
+const ObjectID = require('bson-objectid').default;
+const InMemoryMilestoneRepository = require('../../../../../core/server/services/milestones/InMemoryMilestoneRepository');
+const Milestone = require('../../../../../core/server/services/milestones/Milestone');
 const DomainEvents = require('@tryghost/domain-events');
 const sinon = require('sinon');
