+++ ghost/core/test/unit/server/services/slack-notifications/SlackNotificationsService.test.js
@@ -2,7 +2,7 @@ const assert = require('assert/strict');
 const sinon = require('sinon');
 const SlackNotificationsService = require('../../../../../core/server/services/slack-notifications/SlackNotificationsService');
 const ObjectId = require('bson-objectid').default;
-const {MilestoneCreatedEvent} = require('@tryghost/milestones');
+const MilestoneCreatedEvent = require('../../../../../core/server/services/milestones/MilestoneCreatedEvent');
 const DomainEvents = require('@tryghost/domain-events');
 
 describe('SlackNotificationsService', function () {
