--- ghost/milestones/lib/milestones.js
@@ -1,4 +1,0 @@
-module.exports.InMemoryMilestoneRepository = require('./InMemoryMilestoneRepository');
-module.exports.MilestonesService = require('./MilestonesService');
-module.exports.MilestoneCreatedEvent = require('./MilestoneCreatedEvent');
-module.exports.Milestone = require('./Milestone');
